extract loader-v4-interface crate
