+++ Cargo.toml
@@ -147,6 +147,7 @@ members = [
     "sdk/keccak-hasher",
     "sdk/keypair",
     "sdk/loader-v3-interface",
+    "sdk/loader-v4-interface",
     "sdk/logger",
     "sdk/macro",
     "sdk/message",
@@ -528,6 +529,7 @@ solana-last-restart-slot = { path = "sdk/last-restart-slot", version = "=2.2.0" 
 solana-lattice-hash = { path = "lattice-hash", version = "=2.2.0" }
 solana-ledger = { path = "ledger", version = "=2.2.0" }
 solana-loader-v3-interface = { path = "sdk/loader-v3-interface", version = "=2.2.0" }
+solana-loader-v4-interface = { path = "sdk/loader-v4-interface", version = "=2.2.0" }
 solana-loader-v4-program = { path = "programs/loader-v4", version = "=2.2.0" }
 solana-local-cluster = { path = "local-cluster", version = "=2.2.0" }
 solana-log-collector = { path = "log-collector", version = "=2.2.0" }
